app/requirements.txt
streamlit==1.38.0
pydantic==2.10.6
databricks-sdk==0.44.1
openpyxl==3.1.5
databricks-sql-connector==3.4.0